Add overrides to config and use an example
The old code was only ever used for a single vendor, so the config was
always the same. This time the ambition is to reuse the same code but to
be able to use different settings depending on the use case. For now we
only have the demo index.html page, so we add an example usage there.