Add make target config.json.default
This builds a default configuration that is a mix of

* The example configuration
* Data fetched from Alma (funds, reporting codes 1 and 2)
* Data set in script/default_options.jq

The Make tasks are quite involved, since they both fetch data using the
Alma API and then merges data from several places into a single config
file.

A special construct is used to depend on the value of the environment
variable ALMA_AUTH_HEADER_COMMAND, which is evaluated to generate the
required Authorization header for the API call. Now the existence of the
variable is enforced for the API tasks, triggering rebuilds only
if the variable hast changed.