Rework config further
We now use a default config including fields. On merge, default fields
are now overwritten by custom fields, and then default labels, options
and defaults are merged (overlaid) by the custom config.

We build the example config and the alma default config from the default
config. For the example config we also add some new example options and
defaults to a file in src. This means that from now on, the
config.json.example file is a generated file. It is still checked into
git, though. We'll see about that later on.

The adlibris section of the config is so far unused. We'll see to that
too.