Generate a combined bookmarklet
The index.html page will now include a main.js javascript,
where the adlibris specific code and the standalone form are combined
into a single bookmarklet.

Also, the configs are now merged in make before esbuild

Merging through the JavaScript code leaves overwritten fields in the
build output, increasing the size. Since there is an upper limit for how
long the URL for a bookmarklet can be, it makes sense to reduce this.

This means that the JS code only depends on the original sources
(config.json and src/defaultConfig.json) *indirectly*, and that builds
must be initiated through `make`, where the dependency is expressed, to
apply changes in these files.